Forward logger from compress command to decompress command.
When running in memory (to temporary file) compressiond and then showing
the result, the QCMPDecompressCommand is created in
QCMPCompressCommand which results in the LogService logger not being
initialized by the dependency injection. We can fix this behaviour by
forwarding our logger instance, which was provided to us by the
dependency injection.